Add test coverage for optional serialization dependencies
